requirements.txt
flask
flask-cors
groq
gtts
tavily-python
requests
beautifulsoup4
gunicorn
